refactor : FixedBooleanGenerator 내부 static에서 test모듈 내부로 이동
